service/watcher/requirements.txt
pandas==1.4.3
psycopg2==2.9.3
python-dotenv==1.0.0
requests==2.28.1
slack_sdk==3.20.2
SQLAlchemy==1.4.16
slack_sdk
